# public/swagger-public.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Next Schedule API - Integração N8N",
    "description": "API pública para integrações externas (ex: N8N, Zapier) utilizando API Key.",
    "version": "1.0"
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key"
      }
    }
  },
  "security": [],
  "paths": {
...
    "/api/integrations/clinic-persona": {
      "get": {
        "summary": "Get clinic AI persona settings",
        "description": "Retrieves the AI assistant persona configuration for a clinic identified by phone number. This endpoint is used by N8N to configure the AI agent's behavior, tone, rules, and conversation flow.",
        "tags": [
          "Clinic Persona"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "phone",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The clinic's phone number",
            "example": "+5511999999999"
          }
        ],
        "responses": {
          "200": {
            "description": "Clinic persona settings retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "clinicId": {
                      "type": "string",
                      "format": "uuid",
                      "description": "Unique identifier of the clinic"
                    },
                    "clinicName": {
                      "type": "string",
                      "description": "Name of the clinic"
                    },
                    "clinicType": {
                      "type": "string",
                      "description": "Type/niche of the clinic (e.g., \"Odontologia\", \"Psicologia\")"
                    },
                    "persona": {
                      "type": "object",
                      "nullable": true,
                      "description": "AI persona configuration (null if not configured)",
                      "properties": {
                        "assistantTone": {
                          "type": "string",
                          "description": "The tone the assistant should use (e.g., \"Professional\", \"Friendly\")",
                          "nullable": true
                        },
                        "welcomeMessage": {
                          "type": "string",
                          "description": "Welcome message sent to patients",
                          "nullable": true
                        },
                        "rules": {
                          "type": "array",
                          "description": "Behavioral rules the assistant must follow",
                          "items": {
                            "type": "string"
                          }
                        },
                        "appointmentFlow": {
                          "type": "array",
                          "description": "Steps the assistant should follow when scheduling appointments",
                          "items": {
                            "type": "string"
                          }
                        },
                        "forbiddenTopics": {
                          "type": "array",
                          "description": "Topics the assistant should not discuss",
                          "items": {
                            "type": "string"
                          }
                        },
                        "availability": {
                          "type": "string",
                          "description": "Clinic's availability schedule",
                          "nullable": true
                        },
                        "language": {
                          "type": "string",
                          "description": "Language code (e.g., \"pt-BR\", \"en-US\")",
                          "nullable": true
                        }
                      }
                    }
                  }
                },
                "example": {
                  "clinicId": "550e8400-e29b-41d4-a716-446655440000",
                  "clinicName": "Clínica Exemplo",
                  "clinicType": "Odontologia",
                  "persona": {
                    "assistantTone": "Profissional e amigável",
                    "welcomeMessage": "Olá! Bem-vindo à Clínica Exemplo. Como posso ajudá-lo hoje?",
                    "rules": [
                      "Não fornecer diagnósticos médicos",
                      "Sempre confirmar informações do paciente"
                    ],
                    "appointmentFlow": [
                      "Perguntar o nome do paciente",
                      "Verificar disponibilidade",
                      "Confirmar data e horário"
                    ],
                    "forbiddenTopics": [
                      "Política",
                      "Religião"
                    ],
                    "availability": "Seg-Sex 08:00-18:00",
                    "language": "pt-BR"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Phone number is required",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                },
                "example": {
                  "error": "Phone number is required"
                }
              }
            }
          },
          "404": {
            "description": "Clinic not found with the provided phone number",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                },
                "example": {
                  "error": "Clinic not found"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}